Creato il fil system per le classi che riguardano la ParticleSystem
